+++ src/ByteDev.ValueTypes/EnumTypeHelper.cs
@@ -23,6 +23,20 @@ namespace ByteDev.ValueTypes
         }
 
         /// <summary>
+        /// Returns a list of possible enum values for the given enum type.
+        /// </summary>
+        /// <param name="enumType">The enum type to get as a list.</param>
+        /// <returns>List of possible enum values.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="enumType" /> must be an Enum.</exception>
+        public static IList<Enum> ToList(Type enumType)
+        {
+            return Enum.GetValues(enumType)
+                .Cast<Enum>()
+                .Select(e => e)
+                .ToList();
+        }
+
+        /// <summary>
         /// Returns a list of possible enum value descriptions for the given enum type.
         /// If a enum value has no Description attribute then the enum value as a string is returned.
         /// </summary>
@@ -32,6 +46,21 @@ namespace ByteDev.ValueTypes
         {
             return Enum.GetValues(typeof(TEnum))
                 .Cast<TEnum>()
+                .Select(e => e.GetDescriptionOrName())
+                .ToList();
+        }
+
+        /// <summary>
+        /// Returns a list of possible enum value descriptions for the given enum type.
+        /// If a enum value has no Description attribute then the enum value as a string is returned.
+        /// </summary>
+        /// <param name="enumType">The enum type to get as a list.</param>
+        /// <returns>List of possible enum value's descriptions.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="enumType" /> must be an Enum.</exception>
+        public static IList<string> ToDisplayList(Type enumType)
+        {
+            return Enum.GetValues(enumType)
+                .Cast<Enum>()
                 .Select(e => e.GetDescriptionOrName())
                 .ToList();
         }